Ajuste nos componentes de imagem
